// components/PAGES/DIRECTOR/PropertyDeveloper/VariantA/management/staff-branch/[branchId]/types.d.ts
import type { notificationCardProps } from "@/components/dashboard/types";
interface Stats {
  total: number;
  new_this_month: number;
}

export interface SingleBranchPageData {
  branch_name: string;
  address: string;
  properties: Stats;
  landlords: Stats;
  tenants: Stats;
  vacant_units: Stats;
  expired: Stats;
  invoices: Stats;
  inquiries: Stats;
  complaints: Stats;
  listings: Stats;
  staffs: notificationCardProps["notifications"];
  hasManager: boolean;
  branch_wallet: SubWallet | null;
  picture: string | null;
  recent_transactions: Transactions[];
  transactions: Transactions[];
  receipt_statistics: receipt_statistics | null;
}

interface receipt_statistics {
  total_receipt: string;
  total_paid_receipt: string;
  total_pending_receipt: string;
  percentage_change_paid: number;
  percentage_change_total: number;
  percentage_change_pending: number;
}

interface SubWallet {
  wallet_id: string;
  balance: string;
  escrow_balance: string;
  is_active: string;
  credit_total: number;
  debit_total: number;
  balance_total: string;
  last_week_credit: number;
  last_week_debit: number;
  last_week_balance: number;
}

export interface EditBranchFormData {
  id: string;
  isActive: 1 | 0;
  branch_name: string;
  state: string;
  local_government: string;
  city: string;
  address: string;
  wallet: "yes" | "no";
  description: string;
  picture: string | null;
  hasMoney?: boolean;
  account_name?: string;
  account_number?: string;
  bank_name?: string;
}

export interface Transactions {
  id: number;
  amount: string;
  transaction_type:
    | "debit"
    | "transfer_in"
    | "transfer_out"
    | "withdrawal"
    | "sponsor_listing"
    | "funding";
  reference: string;
  description: string;
  status: string;
  balance_before: string;
  balance_after: string;
  source_name?: string;
  source?: string;
  date: string;
  time: string;
  type: string;
}

export type SingleBranchResponseType = {
  data: {
    branch: {
      id: string;
      branch_name: string;
      is_active: 1 | 0;
      state: string;
      local_government: string;
      city: string;
      branch_address: string;
      picture: string | null;
      // branch_wallet: string; //to be added later
      branch_desc: string;
      landlords_count: number;
      account_name: string;
      account_number: string;
      bank_name: string;
      current_month_landlords_count: number;
      tenants_count: number;
      current_month_tenants_count: number;
      properties_count: number;
      current_month_properties_count: number;
      units_count: number;
      complaints_count: number;
      current_month_complaints_count: number;
      vacant_unit: number;
      vacant_month_unit: number;
      expired_unit: number;
      expired_month_unit: number;
      invoice_count: number;
      current_month_invoice_count: number;
      listing: number;
      listing_month: number;
      staffs_count: number; // do d monthly/this month stuff
      properties_count: number; // do d monthly/this month stuff
      receipt_statistic: receipt_statistics | null;
      staffs: {
        id: string;
        picture: string;
        // user_id: string;
        is_active: 1 | 0;
        title: string | null;
        user_id: string;
        tier: 1 | 2 | 3 | 4 | 5;
        // estate_title: string | null;
        staff_role: "manager" | "staff" | "account officer";
        name: string;
      }[];
    };
    sub_wallet: SubWallet | null;
    recent_transactions: Transactions[];
    transactions: Transactions[];
    manager: {
      id: string;
      is_active: 1 | 0;
      title: string | null;
      estate_title: string | null;
      staff_role: "manager" | "staff" | "account officer";
    }[];
  };
};